+++ src/components/linkcard/LinkCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { Card } from 'antd';
 import { LinkButton } from '../LinkButton';
-import { LIGHT_GREEN, PALE_GREEN, WHITE } from '../../colors';
+import { LIGHT_GREEN, PALE_GREEN, WHITE } from '../../utils/colors';
 import bkg1 from './bkg1.png';
 import bkg2 from './bkg2.png';
 import bkg3 from './bkg3.png';
